+++ src/components/Modal/ModalHeader.tsx
@@ -17,7 +17,7 @@ type TitleProps = {
 ModalHeader.Title = ({ title, subTitle }: TitleProps) => {
   return (
     <ModalHeader>
-      <h5 className="whitespace-pre h5 tex">{title}</h5>
+      <h5 className="whitespace-pre h5">{title}</h5>
       <p className="b1">{subTitle}</p>
     </ModalHeader>
   );
